Seperate peer logic
