+++ infra/network/vlan.go
@@ -1,0 +1,1 @@
+package vlan
